refactor: extract NodeBuilder out of node.go (#3343)
